Moved the Map related files into a subfolder.
