+++ client/src/components/ProTx/components/maps/intervalColorScale.js
